requirements.txt
Django==3.1.6
djangorestframework==3.12.2
drf-yasg==1.20.0
markdown==3.3.3
django-filter==2.4.0
django-admin-interface==0.15.3
Celery==5.0.5
django-celery-results==2.0.1
wcwidth==0.2.5
django-celery-beat==2.2.0
django-redis==4.12.1

